+++ podcasts/Bookmarks/Editing/BookmarkEditing.swift
@@ -5,9 +5,9 @@ import Foundation
 protocol BookmarkEditing: AnyObject {
     var router: BookmarkEditRouter? { get set }
     var analyticsSource: BookmarkAnalyticsSource { get set }
-    var originalTitle: String { get }
 
     func viewDidAppear()
+    func cancel()
 }
 
 extension BookmarkEditing {